Split randstrobes generation from index creation
By moving functions out of index.cpp/.h into new files randstrobes.cpp/.h.

Closes #83
